Move RoomViewHeader into containers folder
