+++ storybook/stories/RoomViewHeader.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { ScrollView, View, StyleSheet } from 'react-native';
 import { HeaderBackButton } from '@react-navigation/stack';
 
-import HeaderComponent from '../../app/views/RoomView/Header/Header';
+import HeaderComponent from '../../app/containers/RoomHeader/Header';
 // import { CustomHeaderButtons, Item } from '../../app/containers/HeaderButton';
 import StoriesSeparator from './StoriesSeparator';
 import { isIOS } from '../../app/utils/deviceInfo';
